// packages/medusa-client/src/admin/models/AdminPostTaxRatesTaxRateProductsParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AdminPostTaxRatesTaxRateProductsParams = {
  /**
   * Which fields should be included in the result.
   */
  fields?: Array<string>;
  /**
   * Which fields should be expanded and retrieved in the result.
   */
  expand?: Array<string>;
};